Some significant design changes
Thanks to u/Triabolical_ and u/alez from r/PrintedCircuitBoards for their
feedback and advice.

1. Switch to 805 resistors and the non-hand-soldered PCB pad footprints
They're not appreciably harder to solder, given that they're only 0.01"
smaller than the 1206 size, and they make routing considerably easier.
The non-hand-soldered size will be a bit trickier, but not all that
much trickier.

2. Move components around
The ESP-12 antenna performs much better when it's open to air on all
sides. It's now at the top of the board, poking off the edge. The switches
and pullup/downs now live right below the ESP-12, the GY521 pins go below
them, and everything else is mostly the same. The ESP-12 smoothing capacitor
now lives directly between the VCC and ground pins.

The overall package is now 0.2mm longer, but the routing is cleaner and
fewer vias are required.