Hotspot mode behavior changes and wifi refactoring
Hotspot mode now tries to connect to the configured wifi network.
It uses the hydrometer name as its mDNS name. If it connects
to a wifi network, it'll try to send updates to a server
in addition to playing local web server.

Moved wifi connection to a separate function for code reuse.